fix: Use long instead of ulong for PinchHitter WebSocket data receive
This is a server for testing. If we are receiving messages over 9.2
zettabytes in length on an incoming WebSocket to the point where we
would need an unsigned long (1.8 yottabytes), we have bigger problems
to worry about.

For reference:
kilo = 1000 (10^3, thousand)
mega = 1000000 (10^6, million)
giga = 1000000000 (10^9, billion)
tera = 1000000000000 (10^12, trillion)
peta = 1000000000000000 (10^15, quadrillion)
exa  = 1000000000000000000 (10^18, pentillion)
zetta = 1000000000000000000000 (10^21, sextillion)
yotta = 1000000000000000000000000 (10^24, septillion)