refactor: Restructure data receive to only log JSON parse errors
Because of the asynchronous nature of the way WebSocket communications
happens in .NET, if a remote end sends a malformed data message,
exceptions thrown during parsing are swallowed by the running data
receiving Task. Rather than throwing an exception here, we will
log the error, and raise an UnknownProtocolMessageReceived event.