feat: Restore use of Channel for Transport message processing
This commit reinstates the use of a Channel for processing incoming
messages in the Transport class. The difference from the previous
implementation is that the class now waits for all pending incoming
messages to be processed before the Disconnect method returns. This
will insure that no sent messages will go unprocessed by the driver.