feat: Make driver sessions completely reusable
A driver session can now be stopped and restarted, either with the
same WebSocket URL, or a different one.

An additional change is that the library will now temporarily process
incoming messages without the defined `type` property. All remote end
implementations have that change in process, and once the stable
channels of the remote end (browser implementations) receive that change,
the TODO comments can be removed. Until that time, code coverage
statistics will be diminished. Those releases are expected by the end of
October 2023.