feat: Allow Transport to surface exceptions (#6)
Since the Transport is running async code, exceptions encountered by event
handlers will be silently swallowed. To be precise, they are raised on the
synchonization context of a different thread than the user's thread. This
change allows those exceptions to be surfaced to the user, and for the user
to specify when those exceptions are propagated (never; at shutdown; or as
soon as possible, usually when the user attempts to execute the next command)