workaround: Temporarily relax protocol compliance
Creating BiDi-only sessions in Firefox has a number of bugs in it where
the browser does not conform to the WebDriver BiDi protocol. This commit
relaxes the strictness of the protocol temporarily to work around those
bugs. The workarounds are marked with TODO comments, and should be
reversed as soon as Firefox fixes the issues at hand. The current issues
are:

* [New session response does not include required property setWindowRect](https://bugzilla.mozilla.org/show_bug.cgi?id=1916522)
* [New session with empty capabilities returns empty proxy property](https://bugzilla.mozilla.org/show_bug.cgi?id=1916463)