chore: use auto properties (#42)
This change moves most of the property definitions in the library to be defined with auto-implemented properties instead of explicit backing fields. This makes the code a little more compact, and should have no impact on public APIs.

In addition, this change also updates collection initialization to use the more concise collection expression ([]). Again, this should introduce no behavior changes in the library.